# docs/swagger.yaml
basePath: /
definitions:
  controllers.createAppointmentRequest:
    properties:
      comment:
        example: Хочу записаться на консультацию
        type: string
      psychologist_id:
        example: 1
        type: integer
    type: object
  controllers.loginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
  controllers.registerRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
  controllers.resetCodeRequest:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
  controllers.resetPasswordRequest:
    properties:
      code:
        example: "123456"
        type: string
      email:
        example: user@example.com
        type: string
      new_password:
        example: newpassword123
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервис авторизации для психологической платформы
  title: Auth Service API
  version: "1.0"
paths:
  /admin/appointments:
    get:
      description: Возвращает все записи — только для администратора
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Все записи (админ)
      tags:
      - admin
  /admin/appointments/{id}/approve:
    patch:
      description: Меняет статус записи на approved
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID записи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Принять запись (админ)
      tags:
      - admin
  /admin/appointments/{id}/reject:
    patch:
      description: Меняет статус записи на rejected
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID записи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Отклонить запись (админ)
      tags:
      - admin
  /appointments:
    post:
      consumes:
      - application/json
      description: Создаёт запись к психологу для авторизованного пользователя
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные записи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.createAppointmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Записаться к психологу
      tags:
      - appointments
  /appointments/my:
    get:
      description: Возвращает список записей авторизованного пользователя
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Мои записи
      tags:
      - appointments
  /auth/login:
    post:
      consumes:
      - application/json
      description: Возвращает JWT токен при успешной авторизации
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход в аккаунт
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя по email и паролю
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Сбрасывает пароль пользователя по коду из email
      parameters:
      - description: Данные для сброса пароля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Сброс пароля
      tags:
      - auth
  /auth/send-reset-code:
    post:
      consumes:
      - application/json
      description: Отправляет 6-значный код на email для сброса пароля
      parameters:
      - description: Email пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.resetCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Отправка кода восстановления
      tags:
      - auth
  /psychologists:
    get:
      description: Возвращает список всех психологов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Список психологов
      tags:
      - psychologists
  /psychologists/{id}:
    get:
      description: Возвращает информацию о конкретном психологе
      parameters:
      - description: ID психолога
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Информация о психологе
      tags:
      - psychologists
swagger: "2.0"